LPS-119634 Allow passing the getPortals function to the FragmentContent to have a different implementation in MasterLayout and Layout
